refactor(server): extract fs helpers to server/utils/fs.ts
Promote statSafe, readDirSafe, and the realpath-based path-traversal
check (now `resolveWithinRoot`) out of routes/files.ts so other route
files can adopt them. Also exposes a `readTextOrNull` helper for the
journal code.

routes/files.ts now imports the helpers and keeps a thin `resolveSafe`
wrapper that adds the hidden-dir (.git) traversal check on top of
`resolveWithinRoot`.

Refs receptron/mulmoclaude#136
